webapp/requirements.txt
pymongo==2.8
mongoengine==0.9.0
Django==1.7.6
django-admin-bootstrapped==2.4.0
django-bootstrap-form==3.2
pinax-theme-bootstrap==5.7.2
django-user-accounts==1.0.1
eventlog==0.10.0
metron==1.3.5
django-crispy-forms==1.4.0
djangorestframework==3.0.5
django-rest-framework-mongoengine==2.0.2
django-datatables-view==1.12
django-extensions==1.5.2
django-compressor==1.4
django-debug-toolbar==1.3.0
django-debug-toolbar-mongo==0.1.10
